delete all data token user by user id before create

+++ src/main/java/com/example/model/UserTokenResponse.java
@@ -3,13 +3,12 @@ package com.example.model;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Data
 @Builder
 public class UserTokenResponse {
     private String token;
-    private Integer user_id;
+    private Long user_id;
     private LocalDateTime expired_at;
 }
